+++ packages/web-wallet/package.json
@@ -45,6 +45,7 @@
     "globals": "^15.11.0",
     "parcel": "^2.13.3",
     "path": "^0.12.7",
+    "svgo": "^3",
     "tailwindcss": "^4.0.0",
     "typescript": "^5.7.3",
     "typescript-eslint": "^8.11.0"
